day 4: using slicing and "expressions"
- I'm not very happy with the parsing code.
It certainly would be easy with regular expressions, but
this sort of parsing wouldn't be hard to do in C/C++ which
you can easily search from a starting point within a string
and then form a string using a begin/end range of
pointers/iterators. Rust doesn't have the means to form a slice
from an interator, which means asking for indices and forming
slices by hand.
- The validation code worked pretty well. I made good use of the
fact that most control flow "statements" can also be expressions.
I was able to write that and get the answer with one compilation
and one run of the program.
- I'm still confused why I needed "by_ref()"