day 16: using regex and "elimination"
- part1 was straightforward with a regular expression
- I initially tried a depth-first search on field-rule
assignments but that wasn't fast enough. I had assumed there
may have been more than one right answer but after inspecting
the field-rule compatabilities it became clear that the assignment
is unique and thus can be solved quickly. Find the field for which
only one rule is compatible and assign it to that rule, remove that
rule from any other field containing it and repeate. This sort of
reminds of gaussian elimination on triangular matrices