day 22: using recursion
- Not super interesting. Algorithmically the obvious
approach is perfectly fine, but I did have part 2 be
too slow because of I tried to hash two deques as my
record of past rounds. It's much faster to convert
them to flat vectors first even if it means copying.
- I'm curious if a card circular queue fixed at a size
large enough to hold any of the cards would be faster
but didn't investigate.
- Eventual solution runs in < 1s