day 13: part 2
- I explored solutions in Octave using some brute force solutions and
looked for patterns
- Generally the approach is something like this: every pair of buses
will together act like a virtual bus with it's own frequency and offset.
The virtual bus is a like a bus that comes every time the pair of buses
arrive lined up with the right offset. It fairly obvious that after some
offset the virtual bus should arrive with a frequency equal to the LCM
of the individual bus frequency. Now it's necessary to find the phase or
offset, that is, the first timestamp where the alignment occurs. Finding
this is a linear diophantine problem. A carefully written brute force
search is sufficent. The offset of the virtual bus is added to the next
offset. Then virtual buses are built pairwise until the end. The offset
of the last virtual bus is the part 2 solution.